+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 [start]
 
+### Bugfixes
+- VE Engine Sliders with multiple profiles were broken
+- Previous VE settings where bleeding into the new VE Engine
+- Multi-opacity select only show if mouseover condition is active
+- VE Profiles are now weighted by position, added an arrow to move positions of profiles.
+- Minimap Missions button would not load after switching to and from HUD map
+- All fonts across orbit now inherit the Shadow option if selected
+
+Previous Update:
+
 ### New!
 - **Visibility Engine Overhaul**
   - Your current visibility options will be reset.
@@ -21,8 +31,8 @@
 - Slight Performance Hack to Group Frames
 - Minor Adjustments to StatusWidget
 
-*Remember, can use the Spotlight feature and type 'Help' to query things about Orbit.*
+Remember, can use the Spotlight feature and type 'Help' to query things about Orbit.
 
-*Things will only get better with YOUR feedback! Jump on the Discord and let me know your pain points*
+Things will only get better with YOUR feedback! Jump on the Discord and let me know your pain points
 
 [end]